rey/requirements.txt
PySide6
anthropic
python-dotenv
